Managing Cart & Modal State
In this commit, a functionality was added to make the cart modal backdrop disappear when the close button is clicked or when the screen background is clicked. When there are two states, such as the cart modal backdrop, it is visible when the cart button is clicked and invisible when the close button is clicked. In cases where there are more than one state, the states must be manipulated using useState.